Add NotFoundException and BadRequestException

+++ src/Core/LeavePro.CleanArch.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,23 +1,25 @@
 ﻿using AutoMapper;
 using LeavePro.CleanArch.Application.Contracts.Persistence;
+using LeavePro.CleanArch.Application.Exceptions;
 using MediatR;
 
 namespace LeavePro.CleanArch.Application.Features.LeaveType.Commands.DeleteLeaveType;
 
 public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeCommand, Unit>
 {
-    private readonly IMapper _mapper;
     private readonly ILeaveTypeRepository _leaveTypeRepository;
 
-    public DeleteLeaveTypeCommandHandler(IMapper mapper, ILeaveTypeRepository leaveTypeRepository)
+    public DeleteLeaveTypeCommandHandler(ILeaveTypeRepository leaveTypeRepository)
     {
-        _mapper = mapper;
         _leaveTypeRepository = leaveTypeRepository;
     }
 
     public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
     {
-        var leaveTypeToDelete = _mapper.Map<Domain.LeaveType>(request);
+        var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
+
+        if (leaveTypeToDelete == null)
+            throw new NotFoundException(nameof(Domain.LeaveType), request.Id);
 
         await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
